requirements.txt
flask>=3.1.0
flask-smorest>=0.45.0
flask-sqlalchemy>=3.1.0
marshmallow>=3.20.0
flask-jwt-extended>=4.4.0
passlib>=1.7.4
flask-cors>=4.0.0
python-dotenv>=1.0.0